LayoutTests/platform/blackberry/accessibility/aria-labelledby-overrides-label-expected.txt

Shut down computer after  minutes
This tests that if aria-labelledby is used, then label elements are not used

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL text.description should be AXDescription: Shut down computer after 10 minutes. Was .
PASS text.titleUIElement() != null && text.titleUIElement().isValid is false
FAIL labelElement.role should be AXRole: AXStaticText. Was .
PASS successfullyParsed is true

TEST COMPLETE
